refactor: LineAddModal 분리, button type 명시
