+++ src/components/ColorPicker/style.js
@@ -38,6 +38,5 @@ export const Color = styled.div`
       border: 3px solid white;
     `};
   background-color: ${({ color }) => color};
-  box-shadow: rgba(60, 64, 67, 0.3) 0px 1px 2px 0px,
-    rgba(60, 64, 67, 0.15) 0px 1px 3px 1px;
+  box-shadow: var(--shadow-button);
 `;
