+++ src/pages/LineManager/index.js
@@ -1,52 +1,9 @@
 import React from 'react';
-import {
-  Button,
-  PageTemplate,
-  ModalTemplate,
-  Input,
-  Selector,
-  ColorPicker,
-} from '../../components';
-import { COLOR, ROUTE, SIZE } from '../../constants';
+import { Button, PageTemplate } from '../../components';
+import { COLOR, ROUTE } from '../../constants';
 import { useModal } from '../../hooks';
-import { ButtonWrapper, Form, SelectorWrapper } from './style';
-
-const LineAddModal = ({ onClickToClose }) => (
-  <ModalTemplate title={'노선 생성'} onClickToClose={onClickToClose}>
-    <Form>
-      <Input
-        type="text"
-        name="line-name"
-        label="노선 이름"
-        placeholder="노선 이름"
-        size={SIZE.LG}
-      />
-      <SelectorWrapper>
-        <Selector
-          name="line-upStation"
-          defaultOption="상행 종점"
-          options={[]}
-          size={SIZE.LG}
-        />
-        <span>↔️</span>
-        <Selector
-          name="line-downStation"
-          defaultOption="하행 종점"
-          options={[]}
-          size={SIZE.LG}
-        />
-      </SelectorWrapper>
-      <Input
-        type="text"
-        name="line-distance"
-        label="거리"
-        placeholder="거리"
-        size={SIZE.LG}
-      />
-      <ColorPicker pickedColor={'test'} onClickColor={() => {}} />
-    </Form>
-  </ModalTemplate>
-);
+import LineAddModal from './LineAddModal';
+import { ButtonWrapper } from './style';
 
 const LineManager = (props) => {
   const { isModalOpen, openModal, handleClickToClose } = useModal();
@@ -55,7 +12,11 @@ const LineManager = (props) => {
     <>
       <PageTemplate title={ROUTE.LINE_MANAGE.NAME}>
         <ButtonWrapper>
-          <Button backgroundColor={COLOR.AMBER} onClick={openModal}>
+          <Button
+            type="button"
+            backgroundColor={COLOR.AMBER}
+            onClick={openModal}
+          >
             노선 추가
           </Button>
         </ButtonWrapper>
